+++ src/entries/popup/pages/home/Points/Points.tsx
@@ -1,401 +1,258 @@
-/* eslint-disable react/jsx-props-no-spreading */
-import { useQuery } from '@tanstack/react-query';
-import { formatDistanceToNowStrict } from 'date-fns';
-import { MotionProps, motion } from 'framer-motion';
-import { PropsWithChildren } from 'react';
-import { Address } from 'wagmi';
+import { motion, useAnimation } from 'framer-motion';
+import { useEffect } from 'react';
 
-import { metadataClient } from '~/core/graphql';
 import { i18n } from '~/core/languages';
 import { useCurrentAddressStore } from '~/core/state';
-import { copy } from '~/core/utils/copy';
-import { formatDate } from '~/core/utils/formatDate';
-import { formatNumber } from '~/core/utils/formatNumber';
-import { Box, Inline, Separator, Stack, Symbol, Text } from '~/design-system';
-import { BoxProps } from '~/design-system/components/Box/Box';
-import { Skeleton } from '~/design-system/components/Skeleton/Skeleton';
+import { useCurrentThemeStore } from '~/core/state/currentSettings/currentTheme';
+import { useFeatureFlagsStore } from '~/core/state/currentSettings/featureFlags';
+import { useTestnetModeStore } from '~/core/state/currentSettings/testnetMode';
+import { TESTNET_MODE_BAR_HEIGHT } from '~/core/utils/dimensions';
+import { Box, Button, Inset, Stack, Text } from '~/design-system';
 import {
-  SymbolName,
-  linearGradients,
+  backgroundColors,
+  globalColors,
 } from '~/design-system/styles/designTokens';
-import { AddressOrEns } from '~/entries/popup/components/AddressOrEns/AddressorEns';
-import { WalletAvatar } from '~/entries/popup/components/WalletAvatar/WalletAvatar';
 
-function Card({
-  children,
-  ...props
-}: PropsWithChildren<BoxProps & MotionProps>) {
+import { ICON_SIZE } from '../../../components/Tabs/TabBar';
+import PointsSelectedIcon from '../../../components/Tabs/TabIcons/PointsSelected';
+import { useAvatar } from '../../../hooks/useAvatar';
+import { useCoolMode } from '../../../hooks/useCoolMode';
+import { useRainbowNavigate } from '../../../hooks/useRainbowNavigate';
+import { ROUTES } from '../../../urls';
+
+import { PointsDashboard } from './PointsDashboard';
+import { usePoints } from './usePoints';
+
+const animationSteps = {
+  one: {
+    scale: 0.9,
+    rotate: -372,
+    x: 0,
+    y: 10.5,
+    transition: { duration: 0.5, ease: [0.2, 0, 0, 1] },
+  },
+  two: {
+    scale: 1.3,
+    rotate: -12,
+    x: 0,
+    y: -16,
+    transition: { duration: 2.5, ease: [0.05, 0.7, 0.1, 1.0] },
+  },
+  three: {
+    scale: 1.2,
+    rotate: -4,
+    x: 4,
+    y: -4,
+    transition: { duration: 2, ease: [0.2, 0, 0, 1] },
+  },
+  four: {
+    scale: 1.1,
+    rotate: -12,
+    x: 0,
+    y: 8,
+    transition: { duration: 2, ease: [0.2, 0, 0, 1] },
+  },
+  five: {
+    scale: 0.9,
+    rotate: -12,
+    x: 0,
+    y: 10.5,
+    transition: { duration: 0.5, ease: [0.2, 0, 0, 1] },
+  },
+  six: {
+    scale: 1.3,
+    rotate: -372,
+    x: 0,
+    y: -16,
+    transition: { duration: 2.5, ease: [0.05, 0.7, 0.1, 1.0] },
+  },
+  seven: {
+    scale: 1.2,
+    rotate: -380,
+    x: -4,
+    y: -4,
+    transition: { duration: 2, ease: [0.2, 0, 0, 1] },
+  },
+  eight: {
+    scale: 1.1,
+    rotate: -372,
+    x: 0,
+    y: 8,
+    transition: { duration: 2, ease: [0.2, 0, 0, 1] },
+  },
+};
+
+const PointsContentPlaceholder = () => {
+  return (
+    <Inset bottom="10px" horizontal="80px">
+      <Stack space="16px">
+        <Text
+          align="center"
+          size="20pt"
+          weight="semibold"
+          color="labelTertiary"
+        >
+          {i18n.t('points.coming_soon_header')}
+        </Text>
+        <Text
+          align="center"
+          color="labelQuaternary"
+          size="12pt"
+          weight="medium"
+        >
+          {i18n.t('points.coming_soon_description')}
+        </Text>
+      </Stack>
+    </Inset>
+  );
+};
+
+const PointsContent = () => {
+  const navigate = useRainbowNavigate();
+  return (
+    <Stack alignHorizontal="center" space="16px">
+      <Inset bottom="10px" horizontal="80px">
+        <Stack space="16px">
+          <Text
+            align="center"
+            size="20pt"
+            weight="semibold"
+            color="labelTertiary"
+          >
+            {i18n.t('points.header')}
+          </Text>
+          <Text
+            align="center"
+            color="labelQuaternary"
+            size="12pt"
+            weight="medium"
+          >
+            {i18n.t('points.description')}
+          </Text>
+        </Stack>
+      </Inset>
+      <Button
+        onClick={() =>
+          navigate(ROUTES.POINTS_ONBOARDING, {
+            state: { skipTransitionOnRoute: ROUTES.HOME },
+          })
+        }
+        color="accent"
+        height="36px"
+        variant="raised"
+      >
+        {i18n.t('points.get_started')}
+      </Button>
+      <Button
+        onClick={() =>
+          navigate(ROUTES.POINTS_REFERRAL, {
+            state: { skipTransitionOnRoute: ROUTES.HOME },
+          })
+        }
+        color="accent"
+        height="36px"
+        variant="tinted"
+      >
+        {i18n.t('points.use_referral_code')}
+      </Button>
+    </Stack>
+  );
+};
+
+function ClaimYourPoints() {
+  const ref = useCoolMode({ emojis: ['🌈', '🎰'] });
+  const { currentAddress } = useCurrentAddressStore();
+  const { data: avatar } = useAvatar({ addressOrName: currentAddress });
+  const { currentTheme } = useCurrentThemeStore();
+  const { testnetMode } = useTestnetModeStore();
+  const { featureFlags } = useFeatureFlagsStore();
+
+  const controls = useAnimation();
+
+  useEffect(() => {
+    const sequenceAnimations = async () => {
+      await controls.start('one');
+      await controls.start('two');
+      await controls.start('three');
+      await controls.start('four');
+      await controls.start('five');
+      await controls.start('six');
+      await controls.start('seven');
+      await controls.start('eight');
+      sequenceAnimations();
+    };
+    sequenceAnimations();
+  }, [controls]);
+
   return (
     <Box
-      as={motion.div}
+      alignItems="center"
       display="flex"
       flexDirection="column"
-      justifyContent="center"
-      gap="12px"
-      paddingVertical="16px"
-      paddingHorizontal="18px"
-      borderRadius="16px"
-      background="surfaceSecondaryElevated"
+      justifyContent="flex-start"
+      marginTop="-20px"
+      paddingTop={featureFlags.points ? '40px' : '80px'}
+      ref={ref}
+      style={{ height: 336 - (testnetMode ? TESTNET_MODE_BAR_HEIGHT : 0) }}
       width="full"
-      boxShadow="12px surfaceSecondaryElevated"
-      {...props}
     >
-      {children}
+      <Stack space="14px">
+        <Box
+          alignItems="center"
+          display="flex"
+          justifyContent="center"
+          style={{
+            transform: 'translateY(-4px)',
+          }}
+        >
+          <Box
+            as={motion.div}
+            animate={controls}
+            alignItems="center"
+            display="flex"
+            initial={{ scale: 1.1, rotate: -372, x: 0, y: 8 }}
+            justifyContent="center"
+            key="pointsAnimation"
+            style={{
+              height: 28,
+              width: 28,
+              willChange: 'transform',
+            }}
+            variants={animationSteps}
+          >
+            <Box
+              position="relative"
+              style={{
+                height: ICON_SIZE,
+                transform: 'scale(0.5)',
+                transformOrigin: 'top left',
+                width: ICON_SIZE,
+                willChange: 'transform',
+              }}
+            >
+              <PointsSelectedIcon
+                accentColor={avatar?.color || globalColors.blue50}
+                colorMatrixValues={null}
+                tintBackdrop={
+                  currentTheme === 'dark'
+                    ? backgroundColors.surfacePrimaryElevated.dark.color
+                    : backgroundColors.surfacePrimaryElevated.light.color
+                }
+                tintOpacity={currentTheme === 'dark' ? 0.2 : 0}
+              />
+            </Box>
+          </Box>
+        </Box>
+        {featureFlags.points ? <PointsContent /> : <PointsContentPlaceholder />}
+      </Stack>
     </Box>
   );
 }
 
-const LeaderboardPositionNumberDisplay = ({
-  position,
-  children,
-}: PropsWithChildren<{ position: number }>) => {
-  const medal = (['🥇', '🥈', '🥉'] as const)[position - 1];
-
-  if (medal)
-    return (
-      <Inline wrap={false} space="8px" alignVertical="center">
-        <Text
-          size="12pt"
-          weight="bold"
-          webkitBackgroundClip="text"
-          background={medal}
-          color="transparent"
-        >
-          {children}
-        </Text>
-        <Text size="16pt" weight="bold">
-          {medal}
-        </Text>
-      </Inline>
-    );
-
-  if (position > 3 && position <= 50)
-    return (
-      <Inline wrap={false} space="8px" alignVertical="center">
-        <Text size="12pt" weight="bold" color="labelTertiary">
-          {children}
-        </Text>
-        <Symbol
-          symbol={`${position}.circle` as SymbolName}
-          size={14}
-          color="labelTertiary"
-          weight="bold"
-        />
-      </Inline>
-    );
-
-  return (
-    <Inline wrap={false} space="8px" alignVertical="center">
-      <Text size="12pt" weight="bold" color="labelTertiary">
-        {children}
-      </Text>
-      <Text size="10pt" weight="bold" color="labelTertiary">
-        #{position}
-      </Text>
-    </Inline>
-  );
-};
-
-const usePoints = (address: Address) => {
-  return useQuery({
-    queryKey: ['points', address],
-    queryFn: async () => {
-      const result = await metadataClient.points({ address });
-      return result.points;
-    },
-  });
-};
-
-function Leaderboard() {
-  const { currentAddress } = useCurrentAddressStore();
-  const { data } = usePoints(currentAddress);
-  if (!data) return null;
-
-  const { user, leaderboard } = data;
-
-  return (
-    <Stack gap="16px">
-      <Text size="16pt" weight="bold">
-        {i18n.t('points.leaderboard')}
-      </Text>
-
-      <Card
-        paddingVertical="12px"
-        paddingHorizontal="16px"
-        flexDirection="row"
-        alignItems="center"
-        justifyContent="space-between"
-      >
-        <Inline wrap={false} space="12px" alignVertical="center">
-          <WalletAvatar
-            size={32}
-            addressOrName={currentAddress}
-            emojiSize="16pt"
-          />
-          <AddressOrEns address={currentAddress} size="14pt" weight="bold" />
-        </Inline>
-        <Text
-          size="16pt"
-          weight="bold"
-          color="accent"
-          textShadow="12px accent text"
-        >
-          #{user.stats.position.current}
-        </Text>
-      </Card>
-      <Card paddingVertical="10px" paddingHorizontal="16px">
-        <Stack separator={<Separator color="separatorTertiary" />} space="12px">
-          {leaderboard.accounts
-            ?.slice(0, 10)
-            .map(({ address, earnings }, index) => (
-              <Inline
-                key={address}
-                wrap={false}
-                space="12px"
-                alignVertical="center"
-                alignHorizontal="justify"
-              >
-                <Inline wrap={false} space="12px" alignVertical="center">
-                  <WalletAvatar
-                    size={32}
-                    addressOrName={address}
-                    emojiSize="16pt"
-                  />
-                  <AddressOrEns address={address} size="14pt" weight="bold" />
-                </Inline>
-                <LeaderboardPositionNumberDisplay position={index + 1}>
-                  {formatNumber(earnings.total)}
-                </LeaderboardPositionNumberDisplay>
-              </Inline>
-            ))}
-        </Stack>
-      </Card>
-    </Stack>
-  );
-}
-
-function TextWithMoreInfo({ children }: PropsWithChildren) {
-  return (
-    <Inline wrap={false} space="4px" alignVertical="center">
-      <Text size="14pt" weight="semibold" color="labelSecondary">
-        {children}
-      </Text>
-      {/* commented for now, we'll add in the future */}
-      {/* <Symbol
-        weight="bold"
-        symbol="info.circle.fill"
-        color="labelQuaternary"
-        size={12}
-      /> */}
-    </Inline>
-  );
-}
-
-function ReferralCode() {
-  const { currentAddress } = useCurrentAddressStore();
-  const { data } = usePoints(currentAddress);
-
-  return (
-    <Stack gap="12px">
-      <TextWithMoreInfo>{i18n.t('points.referral_code')}</TextWithMoreInfo>
-
-      <Inline wrap={false} space="12px">
-        {data ? (
-          <>
-            <Card
-              paddingVertical="12px"
-              whileTap={{ scale: 0.98 }}
-              whileFocus={{ scale: 1.02 }}
-              whileHover={{ scale: 1.02 }}
-              onTap={() =>
-                copy({
-                  value: data.user.referralCode,
-                  title: i18n.t('points.copied_referral_code'),
-                  description: data.user.referralCode,
-                })
-              }
-            >
-              <Text size="20pt" weight="bold" align="center">
-                {data.user.referralCode}
-              </Text>
-            </Card>
-
-            <Card
-              paddingVertical="12px"
-              flexDirection="row"
-              alignItems="center"
-              whileTap={{ scale: 0.98 }}
-              whileFocus={{ scale: 1.02 }}
-              whileHover={{ scale: 1.02 }}
-              onTap={() =>
-                copy({
-                  value: `https://rainbow.me/points?ref=${data.user.referralCode}`,
-                  title: i18n.t('points.copied_referral_link'),
-                  description: `rainbow.me/points?ref=${data.user.referralCode}`,
-                })
-              }
-            >
-              <Symbol
-                symbol="square.on.square"
-                color="accent"
-                filter="shadow 12px accent"
-                weight="bold"
-                size={16}
-              />
-              <Text
-                size="16pt"
-                weight="bold"
-                color="accent"
-                textShadow="12px accent text"
-                align="center"
-              >
-                {i18n.t('copy_link')}
-              </Text>
-            </Card>
-          </>
-        ) : (
-          <>
-            <CardSkeleton height="40px" />
-            <CardSkeleton height="40px" />
-          </>
-        )}
-      </Inline>
-
-      <Text size="12pt" weight="medium" color="labelQuaternary">
-        {i18n.t('points.referral_code_description')}
-      </Text>
-    </Stack>
-  );
-}
-
-function CardSkeleton({ height }: { height: string }) {
-  return <Skeleton height={height} width="100%" style={{ borderRadius: 16 }} />;
-}
-
-function YourRankAndNextDrop() {
-  const { currentAddress } = useCurrentAddressStore();
-  const { data } = usePoints(currentAddress);
-
-  if (!data)
-    return (
-      <Inline wrap={false} space="12px">
-        <CardSkeleton height="89px" />
-        <CardSkeleton height="89px" />
-      </Inline>
-    );
-
-  const { meta, leaderboard, user } = data;
-
-  const nextDistribution = new Date(meta.distribution.next * 1000);
-  const nextDistributionIn = formatDistanceToNowStrict(nextDistribution);
-
-  return (
-    <Inline wrap={false} space="12px">
-      <Card>
-        <TextWithMoreInfo>{i18n.t('points.next_drop')}</TextWithMoreInfo>
-
-        <Text size="20pt" weight="bold">
-          {nextDistributionIn}
-        </Text>
-        <Text
-          size="10pt"
-          weight="bold"
-          color="accent"
-          textShadow="12px accent text"
-        >
-          {formatDate(nextDistribution)}
-        </Text>
-      </Card>
-
-      <Card>
-        <TextWithMoreInfo>{i18n.t('points.your_rank')}</TextWithMoreInfo>
-        <Text size="20pt" weight="bold">
-          #{user.stats.position.current}
-        </Text>
-        <Text
-          size="10pt"
-          weight="bold"
-          color="accent"
-          textShadow="12px accent text"
-        >
-          {i18n.t('points.out_of', {
-            total: leaderboard.stats.total_users,
-          })}
-        </Text>
-      </Card>
-    </Inline>
-  );
-}
-
-const mapToRange = (
-  num: number,
-  inputRange = [0, 3_000_000],
-  outputRange = [30, 300],
-) => {
-  const [inMin, inMax] = inputRange;
-  const [outMin, outMax] = outputRange;
-  const result = ((num - inMin) * (outMax - outMin)) / (inMax - inMin) + outMin;
-  return result > outMax ? outMax : result;
-};
-
-function YourPoints() {
-  const { currentAddress } = useCurrentAddressStore();
-  const { data } = usePoints(currentAddress);
-
-  if (!data)
-    return (
-      <Stack space="12px">
-        <Skeleton height="18px" width="90px" />
-        <Skeleton height="10px" width="40px" />
-        <Skeleton height="9px" width="200px" />
-      </Stack>
-    );
-
-  const { leaderboard, user } = data;
-
-  return (
-    <Stack space="12px">
-      <Text size="26pt" weight="heavy">
-        {formatNumber(user.earnings.total)}
-      </Text>
-      <Box
-        as={motion.div}
-        borderRadius="round"
-        initial={{ width: 40 }}
-        transition={{ duration: 1 }}
-        animate={{
-          width: mapToRange(user.earnings.total, [0, 2_000_000], [40, 300]),
-        }}
-        style={{ height: 10, background: linearGradients.points }}
-      />
-      <Text
-        size="12pt"
-        weight="bold"
-        background="points"
-        color="transparent"
-        webkitBackgroundClip="text"
-      >
-        {i18n.t('points.out_of_current_total_points', {
-          total: formatNumber(leaderboard.stats.total_points),
-        })}
-      </Text>
-    </Stack>
-  );
-}
-
 export function Points() {
-  return (
-    <Stack
-      gap="20px"
-      paddingBottom="120px"
-      separator={<Separator color="separatorTertiary" />}
-      paddingHorizontal="20px"
-      width="full"
-    >
-      <YourPoints />
-      <YourRankAndNextDrop />
-      <ReferralCode />
-      <Leaderboard />
-    </Stack>
-  );
+  const { currentAddress } = useCurrentAddressStore();
+  const { error } = usePoints(currentAddress);
+
+  if (error === 'NON_EXISTING_USER') return <ClaimYourPoints />;
+  return <PointsDashboard />;
 }
